Remove the to_regclass guard on cross-schema voided_label_validation reads (#4878)
The release carrying evolution 355 has reached every prod and test server, so
every schema ConfigTable's fan-out can reach has voided_label_validation and
the probe is dead code. Drop schemaHasVoidedValidationArchive and collapse the
three conditionally-built SQL arms — getCityAggregateDataBySchema,
getContributorUserIdsBySchema, and the Owner scorecard — back to plain queries.

The rollout-safety spec was also the only coverage of the lookup-table arms of
LabelTypeSql, whose probe is still live until #5118, so it moves to
ConfigTableLabelTypeShapeSpec rather than being deleted with the guard.
