+++ app/models/utils/ConfigTable.scala
@@ -135,28 +135,11 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
   }
 
   /**
-   * True iff `voided_label_validation` (created by evolution 355, #4842) exists in the given schema.
-   *
-   * Cross-schema queries must not assume another city's schema is at this instance's evolution level: each city is
-   * its own app instance, so on a rolling release there is a window where this instance (new code) queries a schema
-   * that hasn't applied 354 yet — and a parked deployment (schema present, no running app) may never apply it. A
-   * missing relation would fail the whole query, and the callers' `.recover` would then drop the entire city from
-   * the aggregate surfaces. An unmigrated schema's archive contribution is genuinely 0 (no repair has run there), so
-   * the right behavior is to probe first and skip the archive arm, not to fail. `to_regclass` resolves a relation
-   * name to NULL instead of erroring when absent, which makes it the safe probe.
-   *
-   * @param schema The database schema to probe.
-   * @return       DBIO yielding true when the archive table exists in that schema.
-   */
-  private def schemaHasVoidedValidationArchive(schema: String): DBIO[Boolean] =
-    sql"""SELECT to_regclass('"#$schema".voided_label_validation') IS NOT NULL""".as[Boolean].head
-
-  /**
-   * Whether the schema has replaced its label_type lookup table with the label_type enum (373.sql). Same rolling-
-   * release hazard as schemaHasVoidedValidationArchive: until every deployment is past 373, the fan-out SQL below has
-   * to read whichever shape the other city's schema currently has, or that city silently drops out of the aggregate
-   * surfaces. The lookup table's absence is the signal, since 373 drops it in the same transaction that types the
-   * columns. Delete this probe and the lookup-table arms of LabelTypeSql once the release is on every server (#5118).
+   * Whether the schema has replaced its label_type lookup table with the label_type enum (373.sql). Each city is its
+   * own app instance, so until every deployment is past 373 the fan-out SQL below has to read whichever shape the
+   * other city's schema currently has, or that city errors out and silently drops out of the aggregate surfaces. The
+   * lookup table's absence is the signal, since 373 drops it in the same transaction that types the columns. Delete
+   * this probe and the lookup-table arms of LabelTypeSql once the release is on every server (#5118).
    *
    * @param schema The database schema to probe.
    * @return       DBIO yielding true when the schema is on the enum shape.
@@ -218,21 +201,9 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
    * @param schema The database schema name for the target city
    * @return DBIO action that yields AggregateStats
    */
-  def getCityAggregateDataBySchema(schema: String): DBIO[AggregateStats] = schemaHasVoidedValidationArchive(schema)
-    .zip(schemaHasLabelTypeEnum(schema))
-    .flatMap { case (hasVoidedArchive, hasLabelTypeEnum) =>
+  def getCityAggregateDataBySchema(schema: String): DBIO[AggregateStats] = schemaHasLabelTypeEnum(schema)
+    .flatMap { hasLabelTypeEnum =>
       val labelTypeSql = LabelTypeSql(schema, hasLabelTypeEnum)
-      // Work-credit count (#4842): votes voided by the off-target-markers repair are archived in
-      // voided_label_validation, not deleted, so the "how many validations happened" total keeps counting them.
-      // Guarded per-schema because the table may not exist there yet (see schemaHasVoidedValidationArchive).
-      val voidedCountArm =
-        if (hasVoidedArchive) s""" + (
-              SELECT COUNT(*)
-              FROM "$schema".voided_label_validation
-              INNER JOIN "$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
-              WHERE NOT user_stat.excluded
-          )"""
-        else ""
       sql"""
       SELECT km_audited.km_audited AS km_explored,
             km_audited_no_overlap.km_audited_no_overlap AS km_explored_no_overlap,
@@ -274,12 +245,19 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
               AND deleted = FALSE
               AND tutorial = TRUE
       ) AS tutorial_label_counts, (
+          -- Work-credit count (#4842): votes voided by the off-target-markers repair are archived in
+          -- voided_label_validation, not deleted, so the "how many validations happened" total keeps counting them.
           SELECT (
               SELECT COUNT(*)
               FROM "#$schema".label_validation
               INNER JOIN "#$schema".user_stat ON label_validation.user_id = user_stat.user_id
               WHERE NOT user_stat.excluded
-          )#$voidedCountArm AS validation_count
+          ) + (
+              SELECT COUNT(*)
+              FROM "#$schema".voided_label_validation
+              INNER JOIN "#$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
+              WHERE NOT user_stat.excluded
+          ) AS validation_count
       ) AS total_val_count;
     """
         .as[(Double, Double, Int, Int, Int)]
@@ -314,18 +292,8 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
    * @param schema The database schema to query.
    * @return DBIO action yielding the distinct contributor `user_id`s (a `text` column) for this schema.
    */
-  def getContributorUserIdsBySchema(schema: String): DBIO[Seq[String]] = schemaHasVoidedValidationArchive(schema)
-    .flatMap { hasVoidedArchive =>
-      // Votes voided by the #4842 repair still mark their caster as a contributor — the participation happened.
-      // Guarded per-schema because the archive may not exist there yet (see schemaHasVoidedValidationArchive).
-      val voidedUnionArm =
-        if (hasVoidedArchive) s"""UNION
-      SELECT voided_label_validation.user_id
-      FROM "$schema".voided_label_validation
-      INNER JOIN "$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
-      WHERE NOT user_stat.excluded"""
-        else ""
-      sql"""
+  def getContributorUserIdsBySchema(schema: String): DBIO[Seq[String]] = {
+    sql"""
       SELECT label.user_id
       FROM "#$schema".label
       INNER JOIN "#$schema".user_stat ON label.user_id = user_stat.user_id
@@ -340,9 +308,13 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
       FROM "#$schema".label_validation
       INNER JOIN "#$schema".user_stat ON label_validation.user_id = user_stat.user_id
       WHERE NOT user_stat.excluded
-      #$voidedUnionArm;
+      UNION
+      SELECT voided_label_validation.user_id
+      FROM "#$schema".voided_label_validation
+      INNER JOIN "#$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
+      WHERE NOT user_stat.excluded;
     """.as[String]
-    }
+  }
 
   /**
    * Label type statistics for a specific schema from its existing vote counts, with a row for every type.
@@ -501,27 +473,8 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
         );
       """.as[Boolean].head
 
-    // Both archive reads are guarded per-schema because voided_label_validation may not exist there yet (see
-    // schemaHasVoidedValidationArchive); an unmigrated schema contributes 0 voided votes and no extra contributors.
-    def coreQuery(upToDateFilter: String, hasVoidedArchive: Boolean, hasLabelTypeEnum: Boolean) = {
+    def coreQuery(upToDateFilter: String, hasLabelTypeEnum: Boolean) = {
       val labelTypeSql = LabelTypeSql(schema, hasLabelTypeEnum)
-      // Work-credit add-on (#4842): archived voided votes count toward total_validations (activity volume) but
-      // not the agree/disagree verdict columns. The archive is human-only by construction.
-      val voidedValCountsBody =
-        if (hasVoidedArchive) s"""SELECT COUNT(*) AS cnt
-          FROM "$schema".voided_label_validation
-          INNER JOIN "$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
-          WHERE NOT user_stat.excluded"""
-        else "SELECT 0 AS cnt"
-      // Voided votes (#4842) still mark their caster as a contributor; the archive is human-only by construction,
-      // so no AI-role filter is needed.
-      val voidedContributorArm =
-        if (hasVoidedArchive) s"""UNION
-              SELECT voided_label_validation.user_id AS contributor_id
-              FROM "$schema".voided_label_validation
-              INNER JOIN "$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
-              WHERE NOT user_stat.excluded"""
-        else ""
       sql"""
       SELECT total_streets.cnt          AS total_streets,
              audited_streets.cnt        AS audited_streets,
@@ -619,7 +572,12 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
           LEFT  JOIN sidewalk_login.user_role ON label_validation.user_id = user_role.user_id
           WHERE NOT user_stat.excluded AND user_role.role = 'AI'
       ) AS ai_val_counts, (
-          #$voidedValCountsBody
+          -- Work-credit add-on (#4842): archived voided votes count toward total_validations (activity volume) but
+          -- not the agree/disagree verdict columns, whose verdicts they no longer have.
+          SELECT COUNT(*) AS cnt
+          FROM "#$schema".voided_label_validation
+          INNER JOIN "#$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
+          WHERE NOT user_stat.excluded
       ) AS voided_val_counts, (
           SELECT COUNT(DISTINCT street_edge_id) FILTER (WHERE task_end >= NOW() - INTERVAL '7 days')  AS audits_7d,
                  COUNT(DISTINCT street_edge_id) FILTER (WHERE task_end >= NOW() - INTERVAL '30 days') AS audits_30d
@@ -641,7 +599,13 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
               INNER JOIN "#$schema".user_stat ON label_validation.user_id = user_stat.user_id
               LEFT  JOIN sidewalk_login.user_role ON label_validation.user_id = user_role.user_id
               WHERE NOT user_stat.excluded AND user_role.role IS DISTINCT FROM 'AI'
-              #$voidedContributorArm
+              -- Voided votes (#4842) still mark their caster as a contributor. No AI-role filter here, unlike the
+              -- arms above: the archive is human-only by construction.
+              UNION
+              SELECT voided_label_validation.user_id AS contributor_id
+              FROM "#$schema".voided_label_validation
+              INNER JOIN "#$schema".user_stat ON voided_label_validation.user_id = user_stat.user_id
+              WHERE NOT user_stat.excluded
           ) AS contributor_union
       ) AS active_contributors, (
           -- Distinct EXCLUDED (low-quality) users who placed a label — the data-quality "how much got filtered" signal.
@@ -665,13 +629,8 @@ class ConfigTable @Inject() (protected val dbConfigProvider: DatabaseConfigProvi
     // PostGIS (#4376).
     withJitOff(for {
       hasOutdatedImageryCol <- upToDateFilterQuery
-      hasVoidedArchive      <- schemaHasVoidedValidationArchive(schema)
       hasLabelTypeEnum      <- schemaHasLabelTypeEnum(schema)
-      core                  <- coreQuery(
-        if (hasOutdatedImageryCol) "AND audit_task.outdated_imagery = FALSE" else "",
-        hasVoidedArchive,
-        hasLabelTypeEnum
-      )
+      core <- coreQuery(if (hasOutdatedImageryCol) "AND audit_task.outdated_imagery = FALSE" else "", hasLabelTypeEnum)
       byLabelType <- labelTypeStatsBySchema(schema, LabelTypeSql(schema, hasLabelTypeEnum))
       weeklyTrend <- getCityWeeklyTrendBySchema(schema, Some(ScorecardTrendWeeks))
       output      <- getCityContributorOutputBySchema(schema)
